tank_robot/requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PCA9685==1.0.1
Adafruit-PureIO==0.2.3
astroid==1.4.9
asttokens==1.1.13
automationhat==0.1.0
blinker==1.3
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
chardet==2.3.0
click==6.6
colorama==0.3.7
colorzero==1.1
cryptography==1.7.1
dlib==19.19.0
docutils==0.13.1
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==0.12.1
fourletterphat==0.1.0
gpiozero==1.5.0
idna==2.2
imutils==0.5.2
isort==4.2.5
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.8
keyring==10.1
keyrings.alt==1.3
lazy-object-proxy==1.2.2
MarkupSafe==0.23
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.2
mypy==0.470
numpy==1.12.1
oauthlib==2.0.1
pantilthat==0.0.7
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picraft==1.0
piglow==1.2.4
pigpio==1.38
Pillow==4.0.0
psutil==5.6.2
pyasn1==0.1.9
pybase64==0.5.0
pycrypto==2.6.1
Pygments==2.2.0
pyinotify==0.9.6
PyJWT==1.4.2
pylint==1.6.5
pyOpenSSL==16.2.0
pyperclip==1.5.27
pyserial==3.2.1
pyxdg==0.25
pyzmq==18.0.1
rainbowhat==0.1.0
roman==2.0.0
rpi-ws281x==4.1.0
RPi.GPIO==0.6.5
rplidar==0.9.2
rplidar-roboticia==0.9.5
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
sense-emu==1.1
sense-hat==2.2.0
simplejson==3.10.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
thonny==3.1.0
thonny-pi==1.1
touchphat==0.0.1
twython==3.4.0
typed-ast==0.6.3
unicornhathd==0.0.4
urllib3==1.19.1
Werkzeug==0.11.15
wrapt==1.9.0
zmq==0.0.0
